Centralized all maps fragments to extend a parent framgent and only implement what they have to.
