+++ app/src/main/java/com/example/suberduberuber/Activities/DashboardActivity.java
@@ -26,6 +26,7 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
+import android.location.Location;
 import android.location.LocationManager;
 import android.os.Bundle;
 import android.util.Log;
@@ -34,6 +35,7 @@ import android.view.MenuItem;
 import android.widget.Toast;
 
 import com.example.suberduberuber.Clients.UserClient;
+import com.example.suberduberuber.Fragments.SelectOriginFragment;
 import com.example.suberduberuber.Models.DroppedPinPlace;
 import com.example.suberduberuber.Models.Rider;
 import com.example.suberduberuber.Models.User;
@@ -43,10 +45,14 @@ import com.example.suberduberuber.Repositories.RequestRepository;
 import com.example.suberduberuber.Repositories.UserLocationRepository;
 import com.example.suberduberuber.Repositories.UserRepository;
 import com.example.suberduberuber.Services.LocationService;
+import com.example.suberduberuber.Services.PermissionsService;
 import com.example.suberduberuber.ViewModels.AuthViewModel;
 import com.example.suberduberuber.ViewModels.GetRideViewModel;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GoogleApiAvailability;
+import com.google.android.gms.location.FusedLocationProviderClient;
+import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.model.Dash;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -91,10 +97,11 @@ abstract class DashboardActivity extends AppCompatActivity {
     private DrawerLayout drawerLayout;
     private NavigationView navigationView;
 
-    public boolean locationPermissionGranted = false;
     public static final int PERMISSIONS_REQUEST_ENABLE_GPS = 9002;
-    public static final int PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION = 9003;
-    public static final int ERROR_DIALOG_REQUEST = 9001;
+
+    private PermissionsService permissionsService;
+    private FusedLocationProviderClient fusedLocationProviderClient;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -103,6 +110,8 @@ abstract class DashboardActivity extends AppCompatActivity {
 
         myAuth = FirebaseAuth.getInstance();
 
+        fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(this);
+
         userRepository = new UserRepository();
         requestRepository = new RequestRepository();
         userLocationRepository = new UserLocationRepository();
@@ -110,6 +119,8 @@ abstract class DashboardActivity extends AppCompatActivity {
         navController = Navigation.findNavController(this, getNavHostId());
 
         navigationView = findViewById(R.id.nav_view);
+
+        permissionsService = new PermissionsService(this, this);
 
         drawerLayout = findViewById(getDrawerLayoutId());
         new AppBarConfiguration.Builder(navController.getGraph())
@@ -137,11 +148,11 @@ abstract class DashboardActivity extends AppCompatActivity {
         }
         switch (requestCode) {
             case PERMISSIONS_REQUEST_ENABLE_GPS: {
-                if(locationPermissionGranted){
+                if(permissionsService.locationPermissionGranted){
                     getUserDetails();
                 }
                 else{
-                    getLocationPermission();
+                    permissionsService.getLocationPermission();
                 }
             }
         }
@@ -276,31 +287,47 @@ abstract class DashboardActivity extends AppCompatActivity {
 
     private void getDeviceLocation() {
         if (ContextCompat.checkSelfPermission(this.getApplicationContext(), ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
-            SimpleLocation location = new SimpleLocation(this);
-            double latitude = location.getLatitude();
-            double longitude = location.getLongitude();
-            LatLng deviceLatLng = new LatLng(latitude, longitude);
-            userGeoPoint = new GeoPoint(latitude, longitude);
+            fusedLocationProviderClient.getLastLocation().addOnCompleteListener(new OnCompleteListener<Location>() {
+                @Override
+                public void onComplete(@NonNull Task<Location> task) {
+                    double latitude = task.getResult().getLatitude();
+                    double longitude = task.getResult().getLongitude();
+                    userGeoPoint = new GeoPoint(latitude, longitude);
 
-            userLocation.setGeoPoint(userGeoPoint);
-            userLocation.setTimeStamp(null);
-            saveUserLocation();
-            startLocationService();
+                    userLocation.setGeoPoint(userGeoPoint);
+                    userLocation.setTimeStamp(null);
+                    saveUserLocation();
+                    startLocationService();
+                }
+            });
         }
         else {
-            getLocationPermission();
+            permissionsService.getLocationPermission();
         }
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        if(checkMapServices()) {
-            if(!locationPermissionGranted) {
-                getLocationPermission();
+        if(permissionsService.checkMapServices()) {
+            if(!permissionsService.locationPermissionGranted) {
+                permissionsService.getLocationPermission();
+                getUserDetails();
             }
             else {
                 getUserDetails();
+            }
+        }
+    }
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        permissionsService.locationPermissionGranted = false;
+        switch (requestCode) {
+            case PermissionsService.PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION: {
+                if (grantResults.length > 0
+                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    permissionsService.locationPermissionGranted = true;
+                }
             }
         }
     }
@@ -328,106 +355,5 @@ abstract class DashboardActivity extends AppCompatActivity {
         Log.d(TAG, "isLocationServiceRunning: location service is not running.");
         return false;
     }
-
-
-
-
-
-
-
-    // ALL METHODS BELOW ARE FROM GITHUB
-    // USER: mitchtabian
-    // URL: https://gist.github.com/mitchtabian/2b9a3dffbfdc565b81f8d26b25d059bf
-    // Code is to ask user for permission to use location services for maps
-    private boolean checkMapServices(){
-        if(isServicesOK()){
-            if(isMapsEnabled()){
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private void buildAlertMessageNoGps() {
-        final AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        builder.setMessage("This application requires GPS to work properly, do you want to enable it?")
-                .setCancelable(false)
-                .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
-                    public void onClick(@SuppressWarnings("unused") final DialogInterface dialog, @SuppressWarnings("unused") final int id) {
-                        Intent enableGpsIntent = new Intent(android.provider.Settings.ACTION_LOCATION_SOURCE_SETTINGS);
-                        startActivityForResult(enableGpsIntent, PERMISSIONS_REQUEST_ENABLE_GPS);
-                    }
-                });
-        final AlertDialog alert = builder.create();
-        alert.show();
-    }
-
-    public boolean isMapsEnabled(){
-        final LocationManager manager = (LocationManager) getSystemService( Context.LOCATION_SERVICE );
-
-        if ( !manager.isProviderEnabled( LocationManager.GPS_PROVIDER ) ) {
-            buildAlertMessageNoGps();
-            return false;
-        }
-        return true;
-    }
-
-    private void getLocationPermission() {
-        /*
-         * Request location permission, so that we can get the location of the
-         * device. The result of the permission request is handled by a callback,
-         * onRequestPermissionsResult.
-         */
-        if (ContextCompat.checkSelfPermission(this.getApplicationContext(),
-                android.Manifest.permission.ACCESS_FINE_LOCATION)
-                == PackageManager.PERMISSION_GRANTED) {
-            locationPermissionGranted = true;
-            getUserDetails();
-        } else {
-            ActivityCompat.requestPermissions(this,
-                    new String[]{android.Manifest.permission.ACCESS_FINE_LOCATION},
-                    PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION);
-        }
-    }
-
-    public boolean isServicesOK(){
-        Log.d(TAG, "isServicesOK: checking google services version");
-
-        int available = GoogleApiAvailability.getInstance().isGooglePlayServicesAvailable(DashboardActivity.this);
-
-        if(available == ConnectionResult.SUCCESS){
-            //everything is fine and the user can make map requests
-            Log.d(TAG, "isServicesOK: Google Play Services is working");
-            return true;
-        }
-        else if(GoogleApiAvailability.getInstance().isUserResolvableError(available)){
-            //an error occured but we can resolve it
-            Log.d(TAG, "isServicesOK: an error occured but we can fix it");
-            Dialog dialog = GoogleApiAvailability.getInstance().getErrorDialog(DashboardActivity.this, available, ERROR_DIALOG_REQUEST);
-            dialog.show();
-        }else{
-            Toast.makeText(this, "You can't make map requests", Toast.LENGTH_SHORT).show();
-        }
-        return false;
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode,
-                                           @NonNull String permissions[],
-                                           @NonNull int[] grantResults) {
-        locationPermissionGranted = false;
-        switch (requestCode) {
-            case PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION: {
-                // If request is cancelled, the result arrays are empty.
-                if (grantResults.length > 0
-                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    locationPermissionGranted = true;
-                }
-            }
-        }
-    }
-
-    // ENDS METHODS FROM GIT HUB FOR LOCATION SERVICES
-
 }
