+++ app/src/main/java/com/example/suberduberuber/Activities/MainActivity.java
@@ -34,13 +34,6 @@ public class MainActivity extends AppCompatActivity {
 
     private FirebaseAuth myAuth = FirebaseAuth.getInstance();
     private AuthViewModel authViewModel;
-
-    public boolean locationPermissionGranted = false;
-    public static final int PERMISSIONS_REQUEST_ENABLE_GPS = 9002;
-    public static final int PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION = 9003;
-    public static final int ERROR_DIALOG_REQUEST = 9001;
-    public static final String ERROR_TAG = "Login Activity";
-
     private ProgressBar spinner;
 
     @Override
