+++ app/src/main/res/layout/fragment_confirm_route.xml
@@ -18,7 +18,7 @@
         android:orientation="vertical">
 
         <com.google.android.gms.maps.MapView
-            android:id="@+id/route_map"
+            android:id="@+id/google_map"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             map:uiZoomGestures="false"
