+++ app/src/main/res/layout/fragment_driver_navigation.xml
@@ -10,7 +10,7 @@
 
 
     <com.google.android.gms.maps.MapView
-        android:id="@+id/route_map"
+        android:id="@+id/google_map"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_weight="1"
