+++ app/src/main/res/layout/fragment_ride_pending.xml
@@ -10,7 +10,7 @@
     android:orientation="vertical">
 
     <com.google.android.gms.maps.MapView
-        android:id="@+id/ride_request_live_route"
+        android:id="@+id/google_map"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintBottom_toTopOf="@+id/linearLayout2"
@@ -27,7 +27,7 @@
         android:layout_height="wrap_content"
         android:text="Ride Pending"
         android:textAppearance="@style/TextAppearance.AppCompat.Display1"
-        app:layout_constraintBottom_toBottomOf="@+id/ride_request_live_route"
+        app:layout_constraintBottom_toBottomOf="@+id/google_map"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.074"
         app:layout_constraintStart_toStartOf="parent"
